feat(ts);: rename files & update configs
renamed a lot of files and added some new dependencies
