+++ .eslintrc.json
@@ -6,11 +6,14 @@
     },
     "extends": [
         "airbnb-base",
-        "plugin:jsdoc/recommended"
+        "plugin:jsdoc/recommended",
+        "plugin:@typescript-eslint/recommended"
     ],
     "plugins": [
-        "jsdoc"
+        "jsdoc",
+        "@typescript-eslint"
     ],
+    "parser": "@typescript-eslint/parser",
     "parserOptions": {
         "ecmaVersion": 2018,
         "sourceType": "module"
