--- assets/js/users/login.js
@@ -1,2 +1,0 @@
-// Importing this file, will auto-show the Consent popup
-import 'JS/lib/cookie';
