+++ assets/js/users/login.ts
@@ -1,0 +1,2 @@
+// Importing this file, will auto-show the Consent popup
+import 'JS/lib/cookie';
