--- assets/js/welcome.js
@@ -1,1 +1,0 @@
-import 'CSS/welcome.scss';
