+++ assets/js/welcome.ts
@@ -1,0 +1,1 @@
+import 'CSS/welcome.scss';
