--- jsconfig.service-worker.json
@@ -1,7 +1,0 @@
-{
-    "extends": "./jsconfig",
-    "compilerOptions": {
-        "lib": ["WebWorker", "ES2015"]
-    },
-    "files": ["sw.js"]
-}
