+++ package.json
@@ -24,6 +24,7 @@
         "regenerator-runtime": "^0.13.2",
         "sass": "^1.32.8",
         "sass-loader": "^10.1.1",
+        "ts-loader": "8.0.1",
         "twemoji": "^13.0.2",
         "url-loader": "^4.1.0",
         "uuid": "^8.3.2",
@@ -32,6 +33,11 @@
         "yaml-front-matter": "^4.1.1"
     },
     "devDependencies": {
+        "@types/jquery": "^3.5.5",
+        "@types/materialize-css": "^1.0.9",
+        "@types/uuid": "^8.3.0",
+        "@typescript-eslint/eslint-plugin": "^4.24.0",
+        "@typescript-eslint/parser": "^4.24.0",
         "eslint": "^7.21.0",
         "eslint-config-airbnb-base": "^14.2.0",
         "eslint-import-resolver-alias": "^1.1.2",
@@ -40,7 +46,8 @@
         "eslint-plugin-jsdoc": "^30.1.0",
         "jshint": "^2.12.0",
         "stylelint": "^13.12.0",
-        "stylelint-config-standard": "^20.0.0"
+        "stylelint-config-standard": "^20.0.0",
+        "typescript": "^4.2.4"
     },
     "license": "SEE LICENSE IN 'LICENSE'",
     "private": false,
