+++ webpack.config.js
@@ -38,13 +38,13 @@ Encore
      * Each entry will result in one JavaScript file (e.g. app.js)
      * and one CSS file (e.g. app.css) if your JavaScript imports CSS.
      */
-    .addEntry('app', './assets/js/app.js')
-    .addEntry('welcome', './assets/js/welcome.js')
-    .addEntry('notes', './assets/js/notes/main.js')
-    .addEntry('user-admin', './assets/js/users/admin.js')
-    .addEntry('user-create', './assets/js/users/create.js')
-    .addEntry('user-login', './assets/js/users/login.js')
-    .addEntry('user-edit', './assets/js/users/edit.js')
+    .addEntry('app', './assets/js/app.ts')
+    .addEntry('welcome', './assets/js/welcome.ts')
+    .addEntry('notes', './assets/js/notes/main.ts')
+    .addEntry('user-admin', './assets/js/users/admin.ts')
+    .addEntry('user-create', './assets/js/users/create.ts')
+    .addEntry('user-login', './assets/js/users/login.ts')
+    .addEntry('user-edit', './assets/js/users/edit.ts')
 
     // When enabled, Webpack "splits" your files into smaller pieces for greater optimization.
     .splitEntryChunks()
@@ -92,7 +92,7 @@ Encore
     .addLoader({ test: /\.md$/, loader: 'raw-loader' })
 
     // uncomment if you use TypeScript
-    //.enableTypeScriptLoader()
+    .enableTypeScriptLoader()
 
     // uncomment to get integrity="..." attributes on your script & link tags
     // requires WebpackEncoreBundle 1.4 or higher
